+++ commands/browse_test.go
@@ -5,7 +5,6 @@ import (
 	"os"
 	"testing"
 
-	"github.com/lighttiger2505/lab/cmd"
 	"github.com/lighttiger2505/lab/config"
 	"github.com/lighttiger2505/lab/git"
 	"github.com/lighttiger2505/lab/gitlab"
@@ -27,6 +26,12 @@ var mockGitClient = &git.MockClient{
 	},
 }
 
+type MockURLOpener struct{}
+
+func (m *MockURLOpener) Open(url string) error {
+	return nil
+}
+
 func TestBrowseCommandRun(t *testing.T) {
 	mockUI := ui.NewMockUi()
 
@@ -45,51 +50,10 @@ func TestBrowseCommandRun(t *testing.T) {
 		Ui:        mockUI,
 		Provider:  provider,
 		GitClient: mockGitClient,
-		Cmd:       cmd.NewMockCmd("browse"),
+		Opener:    &MockURLOpener{},
 	}
 	args := []string{}
 	if code := c.Run(args); code != 0 {
 		t.Fatalf("wrong exit code. errors: \n%s", mockUI.ErrorWriter.String())
-	}
-}
-
-var gitlabRemoteTest = &git.RemoteInfo{
-	Domain:     "domain",
-	NameSpace:  "namespace",
-	Repository: "project",
-}
-
-type getUrlByUserSpecificTest struct {
-	gitlabRemote *git.RemoteInfo
-	args         []string
-	domain       string
-	url          string
-	err          error
-}
-
-type getUrlByRemoteTest struct {
-	gitlabRemote *git.RemoteInfo
-	args         []string
-	branch       string
-	url          string
-	err          error
-}
-
-type searchBrowserLauncherTest struct {
-	goos    string
-	browser string
-}
-
-var searchBrowserLauncherTests = []searchBrowserLauncherTest{
-	{goos: "darwin", browser: "open"},
-	{goos: "windows", browser: "cmd /c start"},
-}
-
-func TestSearchBrowserLauncher(t *testing.T) {
-	for i, test := range searchBrowserLauncherTests {
-		browser := searchBrowserLauncher(test.goos)
-		if test.browser != browser {
-			t.Errorf("#%d: bad return value \nwant %#v \ngot  %#v", i, test.browser, browser)
-		}
 	}
 }
